+++ src/org/labkey/test/components/ui/grids/EditableGrid.java
@@ -1,6 +1,5 @@
 package org.labkey.test.components.ui.grids;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.SystemUtils;
 import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
@@ -368,12 +367,7 @@ public class EditableGrid extends WebDriverComponent<EditableGrid.ElementCache>
             if (useDatePicker)
             {
                 ReactDatePicker datePicker = elementCache().datePicker();
-                datePicker.set("", false);
-                String[] dateParts = getParsedDateParts((String) value);
-                datePicker.set(dateParts[0], false); // use keyboard input to set year and month
-                elementCache().datePickerDateCell(dateParts[1]).click(); // use calendar ui to select day
-                if (!StringUtils.isEmpty(dateParts[2])) // use timepicker to select time
-                    datePicker.clickTime(dateParts[2]);
+                datePicker.select((String) value);
 
                 WebElement inputCell = elementCache().inputCell();
                 inputCell.sendKeys(Keys.RETURN);
@@ -394,34 +388,6 @@ public class EditableGrid extends WebDriverComponent<EditableGrid.ElementCache>
             WebDriverWrapper.waitFor(() -> gridCell.getText().contains(value.toString()),
                     "Value entered into inputCell '" + value + "' did not appear in grid cell.", WAIT_FOR_JAVASCRIPT);
         }
-    }
-
-    /**
-     * Parse "2022-07-11 08:30" to ["2022-07", "11", "8:30 AM"]
-     * @param fullDateStr
-     * @return
-     */
-    private String[] getParsedDateParts(String fullDateStr)
-    {
-        String[] parts = fullDateStr.split(" ") ;
-        String[] dayParts = parts[0].split("-");
-        String yearMon = dayParts[0] + "-" + dayParts[1];
-        String dayPart = dayParts[2];
-        String timePart = parts.length > 1 ? parts[1] : "";
-        if (!StringUtils.isEmpty(timePart))
-        {
-            String[] timeParts = timePart.split(":");
-            String amPM = "AM";
-            int hour = Integer.parseInt(timeParts[0]);
-            if (hour > 12)
-            {
-                amPM = "PM";
-                hour -= 12;
-            }
-            timePart = hour + ":" + timeParts[1] + " " + amPM;
-        }
-
-        return new String[]{yearMon, dayPart, timePart};
     }
 
     /**
@@ -817,16 +783,6 @@ public class EditableGrid extends WebDriverComponent<EditableGrid.ElementCache>
             return new ReactDatePicker.ReactDateInputFinder(getDriver()).withClassName("date-input-cell").find(getComponentElement());
         }
 
-        /**
-         * Return the date cell div of react datepicker
-         * @param day '01', '02', ... '31'
-         * @return
-         */
-        public WebElement datePickerDateCell(String day)
-        {
-            return Locators.datePickerDate(day).findElement(getComponentElement());
-        }
-
     }
 
     protected abstract static class Locators
@@ -851,11 +807,6 @@ public class EditableGrid extends WebDriverComponent<EditableGrid.ElementCache>
         static Locator.XPathLocator itemElement(String text)
         {
             return Locator.tagContainingText("span", text).withClass("btn-primary");
-        }
-
-        static Locator.XPathLocator datePickerDate(String datePart)
-        {
-            return Locator.tagWithClass("div", "react-datepicker__day--0" + datePart);
         }
 
     }
